feat(order): redirect after order submits
A redirect to services&product.html was added to the item.html right
after 2 seconds of the order submission. Before order was submitted and
the page remained as if nothing happened with only an alert with the
message that order was submitted successfully but now it also redirects
the user to the services&product.html to make them see more products and
hopefully make them buy or order more.

I think it will help with overall UX of the site. closes #4